Refactor: Discreminate test file of application interface
