+++ config/app_map.json
@@ -5,7 +5,7 @@
         "Account": "ste",
         "Password": "123456",
         "Local Directory": "/home/pi",
-        "Remote Directory": "C:\\Users\\Administrator\\Documents\\Projects\\AutoRAID"
+        "Remote Directory": "C:\\Users\\STE\\Projects\\AutoRAID"
     },
     {
         "Remote": "192.168.0.128",
