+++ tests/test_amd_desktop/test_application_interface.py
@@ -10,13 +10,19 @@ logging.basicConfig(
     format='%(asctime)s %(levelname)-8s %(message)s', datefmt='%Y-%m-%d %H:%M:%S')
 logger = logging.getLogger(__name__)
 
-LINUX_CMD_TABLE = ({"Command": "pwd",
-                    "Return": {"remote": {0: '/home/test/FW_Generic'},
-                               "local": {0: '/home/pi'}}
+WINDOWS_CMD_TABLE = ({
+                        "Command": "cd",
+                        "Return": {
+                            "remote": {0: "C:\\Users\\STE\\Projects\\AutoRAID"}
+                        }
                     },
-                   {"Command": "uname -r",
-                    "Return": {"remote": {0: '4.10.0-40-generic'},
-                               "local": {0: '6.1.21-v8+'}}
+                   {
+                       "Command": "ver",
+                        "Return": {
+                            "remote": {
+                                0: 'Microsoft Windows [Version 10.0.19045.4780]'
+                            }
+                        }
                     })
 
 class TestApplicationInterface(object):
@@ -26,10 +32,10 @@ class TestApplicationInterface(object):
     #     return api(cmdopt.get('mode'),
     #                cmdopt.get('if_name'),
     #                cmdopt.get('config_file'))
-    @pytest.mark.skip(reason="Current SUT is Windows")
-    @pytest.mark.parametrize("linux_cmd", LINUX_CMD_TABLE)
-    def test_command_line(self, my_app, linux_cmd):
-        list_executed = my_app.command_line(linux_cmd["Command"])
+    # @pytest.mark.skip(reason="Current SUT is Windows")
+    @pytest.mark.parametrize("windows_cmd", WINDOWS_CMD_TABLE)
+    def test_command_line(self, my_app, windows_cmd):
+        list_executed = my_app.command_line(windows_cmd["Command"])
         logger.debug("my_app.command_line = %s, type %s",
                     list_executed, type(list_executed))
-        assert list_executed == linux_cmd["Return"][my_app.mode]
+        assert list_executed == windows_cmd["Return"][my_app.mode]
