+++ tests/test_unit/conftest.py
@@ -4,7 +4,7 @@
 import logging
 import pytest
 
-from unit.application_interface import ApplicationInterface as api
+# from unit.application_interface import ApplicationInterface as api
 
 logging.getLogger("pymongo").setLevel(logging.CRITICAL)
 logging.getLogger('unit.system_under_testing').setLevel(logging.CRITICAL)
@@ -13,11 +13,11 @@ logging.getLogger('unit.ping').setLevel(logging.CRITICAL)
 
 logger = logging.getLogger(__name__)
 
-@pytest.fixture(scope="session", autouse=True)
-def my_app(cmdopt):
-    print('\n\033[32m================== Setup API =================\033[0m')
-    return api(cmdopt.get('mode'), cmdopt.get('if_name'),
-        cmdopt.get('config_file'))
+# @pytest.fixture(scope="session", autouse=True)
+# def my_app(cmdopt):
+#     print('\n\033[32m================== Setup API =================\033[0m')
+#     return api(cmdopt.get('mode'), cmdopt.get('if_name'),
+#         cmdopt.get('config_file'))
 
 @pytest.fixture(scope="session", autouse=True)
 def gitlab_api(request):
